# prevent third-party types in twoslash snippets