@reillylm 2025-01-25
Sorry for the novel; I figured this was an unusual change so it warranted some explanation. However, after writing everything up, I see that we already discussed this a little bit in #3165, so a TL;DR for this change would be "doing the same thing to the Gatsby build that we did to the twoslash tests in #3165" 🙂 

If we wanted to use ES2023 as our `target` instead, there's one snippet who changes, in "Classes," where we [show off overriding properties from `Function`](https://www.typescriptlang.org/docs/handbook/2/classes.html#special-static-names). In the snippet, we want to get a TS2699, but with `target >= es2022`, the compiler sets `useDefineForClassFields` to `true` by default, which makes the error go away. This is the same issue as #3299. If the example is no longer accurate for newer `target`s, I wonder if we should just remove it?