@jakebailey 2025-01-25
CI went from 23 minutes to 13 minutes, awesome.